% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{FAOSYB-package}
\alias{FAOSYB-package}
\title{Package to provide a harmonized framework to deal with
official statistics from different sources.}
\description{
Package to facilitate the use of data from FAO and World Bank
}
\details{
\tabular{ll}{
  Package: \tab FAOSYB \cr
  Type: \tab Package\cr
  Version: \tab 0.1\cr
  Date: \tab 2013-01-03\cr
  License: \tab GPL (>= 2)\cr
  LazyLoad: \tab yes\cr
}
}
\author{
Michael C. J. Kao \email{michael.kao@fao.org}
}
\keyword{package}

