\name{plot_syb}
\alias{plot_syb}
\title{A function for standardised SYB plots}
\usage{
  plot_syb(x, y, group = NULL, type, subset = TRUE, data,
    scale = 1, x_lab = NULL, y_lab = NULL,
    legend_lab = NULL, col_pallete, nCnty = 20)
}
\arguments{
  \item{x}{}

  \item{y}{}

  \item{data}{}

  \item{group}{}

  \item{type}{}

  \item{x_lab}{}

  \item{y_lab}{}

  \item{legend_lab}{}
}
\description{
  A function for standardised SYB plots
}

